backend/requirements.txt
playwright~=1.54.0
bs4~=0.0.2
langchain_google_vertexai~=2.0.27
langchain
langchain_core~=0.3.72
dotenv~=0.9.9
chromadb~=1.0.15
langchain-community
langchain_chroma
python-dotenv~=1.1.1
vertexai
numpy~=2.3.1
requests~=2.32.4
beautifulsoup4~=4.13.4
langchain-text-splitters~=0.3.8
pydantic~=2.11.7
langgraph~=0.6.2
uvicorn~=0.35.0
fastapi~=0.116.1
httpx~=0.28.1
python-jose~=3.5.0
nest-asyncio~=1.6.0
sympy~=1.14.0
ipython~=9.4.0
scikit-learn~=1.7.1
google-cloud-firestore==2.14.0
